Renamed the context menu to distinguish it.
